+++ compiler-core/src/ast/constant.rs
@@ -90,6 +90,12 @@ pub enum Constant<T, RecordTag> {
         /// states.
         extra_information: Option<InvalidExpression>,
     },
+
+    Todo {
+        location: SrcSpan,
+        type_: T,
+        message: Option<Box<Self>>,
+    },
 }
 
 impl TypedConstant {
@@ -105,6 +111,7 @@ impl TypedConstant {
             | Constant::Record { type_, .. }
             | Constant::RecordUpdate { type_, .. }
             | Constant::Var { type_, .. }
+            | Constant::Todo { type_, .. }
             | Constant::Invalid { type_, .. } => type_.clone(),
         }
     }
@@ -118,6 +125,12 @@ impl TypedConstant {
             | Constant::Float { .. }
             | Constant::String { .. }
             | Constant::Invalid { .. } => Located::Constant(self),
+
+            Constant::Todo { message, .. } => message
+                .iter()
+                .find_map(|message| message.find_node(byte_index))
+                .unwrap_or(Located::Constant(self)),
+
             Constant::Var {
                 module: Some((module_alias, location)),
                 constructor: Some(constructor),
@@ -182,6 +195,7 @@ impl TypedConstant {
             | Constant::Tuple { .. }
             | Constant::List { .. }
             | Constant::BitArray { .. }
+            | Constant::Todo { .. }
             | Constant::StringConcatenation { .. }
             | Constant::Invalid { .. } => None,
             Constant::Record {
@@ -207,9 +221,20 @@ impl TypedConstant {
             | Constant::Float { .. }
             | Constant::String { .. } => im::hashset![],
 
-            Constant::List { elements, .. } | Constant::Tuple { elements, .. } => elements
+            Constant::Todo { message, .. } => message
+                .as_ref()
+                .map(|message| message.referenced_variables())
+                .unwrap_or(im::hashset![]),
+
+            Constant::Tuple { elements, .. } => elements
                 .iter()
                 .map(|element| element.referenced_variables())
+                .fold(im::hashset![], im::HashSet::union),
+
+            Constant::List { elements, tail, .. } => elements
+                .iter()
+                .map(|element| element.referenced_variables())
+                .chain(tail.iter().map(|tail| tail.referenced_variables()))
                 .fold(im::hashset![], im::HashSet::union),
 
             Constant::Record { arguments, .. } => arguments
@@ -245,6 +270,19 @@ impl TypedConstant {
 
     pub(crate) fn syntactically_eq(&self, other: &Self) -> bool {
         match (self, other) {
+            (
+                Constant::Todo { message, .. },
+                Constant::Todo {
+                    message: other_message,
+                    ..
+                },
+            ) => match (message, other_message) {
+                (None, None) => true,
+                (Some(_), None) | (None, Some(_)) => false,
+                (Some(message), Some(other_message)) => message.syntactically_eq(other_message),
+            },
+            (Constant::Todo { .. }, _) => false,
+
             (Constant::Int { int_value: n, .. }, Constant::Int { int_value: m, .. }) => n == m,
             (Constant::Int { .. }, _) => false,
 
@@ -273,14 +311,23 @@ impl TypedConstant {
             (Constant::Tuple { .. }, _) => false,
 
             (
-                Constant::List { elements, .. },
+                Constant::List { elements, tail, .. },
                 Constant::List {
                     elements: other_elements,
+                    tail: other_tail,
                     ..
                 },
-            ) => pairwise_all(elements, other_elements, |(one, other)| {
-                one.syntactically_eq(other)
-            }),
+            ) => {
+                let tails_are_equal = match (tail, other_tail) {
+                    (None, None) => true,
+                    (None, Some(_)) | (Some(_), None) => false,
+                    (Some(tail), Some(other_tail)) => tail.syntactically_eq(other_tail),
+                };
+                tails_are_equal
+                    && pairwise_all(elements, other_elements, |(one, other)| {
+                        one.syntactically_eq(other)
+                    })
+            }
             (Constant::List { .. }, _) => false,
 
             (
@@ -385,18 +432,27 @@ impl TypedConstant {
         }
     }
 
+    /// If the constant is a list whose elements are known at compile time, this
+    /// returns a vec of all its elements.
+    /// This can happen with:
+    /// - literal lists: `[1, 2]`
+    /// - literal lists whose tail is also known at compile time:
+    ///     - `[1, 2, ..[3, 4]]`
+    ///     - `[1, 2, ..a_known_module_constant]`
+    ///
     pub fn list_elements(&self) -> Option<Vec<&Self>> {
         match self {
-            Constant::List { elements, tail, .. } => Some(
-                elements
-                    .iter()
-                    .chain(
-                        tail.as_deref()
-                            .and_then(|tail| tail.list_elements())
-                            .unwrap_or_default(),
-                    )
-                    .collect(),
-            ),
+            Constant::List { elements, tail, .. } => {
+                if let Some(tail) = tail {
+                    // There's a tail, if it cannot be known at compile time,
+                    // then this entire list cannot be known at compile time!
+                    let tail_elements = tail.list_elements()?;
+                    Some(elements.iter().chain(tail_elements).collect())
+                } else {
+                    // There's no tail, we just return the elements
+                    Some(elements.iter().collect())
+                }
+            }
             Constant::Var {
                 constructor: Some(constructor),
                 ..
@@ -416,6 +472,7 @@ impl TypedConstant {
             | Constant::BitArray { .. }
             | Constant::StringConcatenation { .. }
             | Constant::Var { .. }
+            | Constant::Todo { .. }
             | Constant::Invalid { .. } => None,
         }
     }
@@ -440,6 +497,7 @@ impl<A, B> Constant<A, B> {
             | Constant::BitArray { location, .. }
             | Constant::Var { location, .. }
             | Constant::Invalid { location, .. }
+            | Constant::Todo { location, .. }
             | Constant::StringConcatenation { location, .. } => *location,
         }
     }
@@ -453,6 +511,7 @@ impl<A, B> Constant<A, B> {
             | Constant::Var { .. } => true,
 
             Constant::Tuple { .. }
+            | Constant::Todo { .. }
             | Constant::List { .. }
             | Constant::Record { .. }
             | Constant::RecordUpdate { .. }