+++ CHANGELOG.md
@@ -4,6 +4,46 @@
 
 ### Compiler
 
+- The inference of record update expressions is now more fault tolerant: if
+  there's an error in the record being updated, the compiler can still able to
+  analyse the fields that are being provided.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
+- It is now possible to use the `todo` keyword in constants, this will result in
+  an helpful error message rather than a syntax error.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
+- The compiler now prints correctly qualified or aliased type names when
+  printing warnings. For example:
+
+  ```gleam
+  import user
+
+  pub fn main() {
+    user.to_string(todo)
+    |> io.println
+  }
+  ```
+
+  Will produce the following warning:
+
+  ```
+  warning: Todo found
+    ┌─ /src/warning/wrn.gleam:4:19
+    │
+  4 │     user.to_string(todo)
+    │                    ^^^^ This code is incomplete
+
+  This code will crash if it is run. Be sure to finish it before
+  running your program.
+
+  Hint: I think its type is `user.User`.
+  ```
+
+  Notice how the type hint is correctly qualified for the module the warning is
+  raised in.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
 ### Build tool
 
 - The `gleam dev` command now accepts the `--no-print-progress` flag. When this
@@ -11,6 +51,33 @@
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
 ### Language server
+
+- The language server can now help with completions when typing a list's tail:
+
+  ```gleam
+  pub fn main() {
+    let things_i_like = ["Gleam", "Ice Cream"]
+    ["Dogs", ..t|]
+    //          ^ Can now suggest a completion for `things_i_like`
+  }
+  ```
+
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
+- The language server can now help with completions when typing a record update:
+
+  ```gleam
+  pub type User {
+    User(name: String, likes: List(String))
+  }
+
+  pub fn set_name(user: User, name: String) -> User {
+    User(..u|)
+    //      ^ Can now suggest a completion for `user`
+  }
+  ```
+
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
 ### Formatter
 
@@ -20,6 +87,32 @@
   or git dependency on Hex when running `gleam update`.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
+- Fixed a bug where the "pattern match on value" code action would generate
+  invalid code when used on a `let` assignment on the right hand side of another
+  `let` assignment.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
+- Fixed a bug where the compiler wouldn't track the minimum required version
+  when using list prepending in constants.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
 - Fixed a bug where the language server wouldn't let one extract record
   constructors as variables.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
+- Fixed a bug where the language server would suggest completions for values
+  from the language's prelude, even though their types were incompatible with
+  the current context.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
+- Fixed a bug where the language server would suggest the "wrap in anonymous"
+  code action even when not hovering directly over a function.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
+- Fixed a bug where the language server would suggest the "wrap in anonymous"
+  code action when hovering over a record update.
+  ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
+
+- Fixed a bug where the compiler would generate invalid code for guards using
+  lists with a tail.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))